remove txFormat
