+++ src/app/ui/util.js
@@ -1,0 +1,23 @@
+var TreeView = require('remix-debugger').ui.TreeView
+var ethJSUtil = require('ethereumjs-util')
+var BN = ethJSUtil.BN
+var remixLib = require('remix-lib')
+var txFormat = remixLib.execution.txFormat
+
+module.exports = {
+  decodeResponseToTreeView: function (response, fnabi) {
+    var treeView = new TreeView({
+      extractData: (item, parent, key) => {
+        var ret = {}
+        if (BN.isBN(item)) {
+          ret.self = item.toString(10)
+          ret.children = []
+        } else {
+          ret = treeView.extractDataDefault(item, parent, key)
+        }
+        return ret
+      }
+    })
+    return treeView.render(txFormat.decodeResponse(response, fnabi))
+  }
+}
